api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
attrs==23.1.0
Automat==22.10.0
bcrypt==4.0.1
binaryornot==0.4.4
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cookiecutter==2.4.0
cryptography==41.0.5
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
incremental==22.10.0
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
keyring==24.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
Pygments==2.16.1
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-okx==0.2.3
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
rsa==4.9
SecretStorage==3.3.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
text-unidecode==1.3
twisted==23.10.0
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
zipp==3.17.0
zope.interface==6.1
loguru